requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
gunicorn>=20.1.0
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
requests>=2.31.0 